+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.6.1
 - Made DynamicTextAssets#dynamicData readonly.
+- Added DynamicTextAssets#getReplaceData() and DynamicTextAssets#getNullableReplacedData().
 # 0.6.0
 - Created DynamicTextAssets class for TextAssets which has placeholders.
 - Added LeveledSkill#paramList and LeveledSkill#simpleParamList.